Controler do usuario é login feitos

+++ BoletimVenus/src/main/java/br/edu/ifpb/pweb2/BoletimVenusSB/BoletimVenus/Repository/EstudanteRepository.java
@@ -3,6 +3,6 @@ package br.edu.ifpb.pweb2.BoletimVenusSB.BoletimVenus.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifpb.pweb2.BoletimVenusSB.BoletimVenus.Model.Estudante;
 
-public interface EstudanteRepository extends JpaRepository<Estudante, Long>{
+public interface EstudanteRepository extends JpaRepository<Estudante, Integer>{
 
 }
